Add alpha to color pickers
Something seems to be wrong with UIElement, if alpha is hidden in the color picker it will sometimes be set to 0. This makes some of the rays invisible

+++ Assets/Scripts/CameraVisualizer.cs
@@ -46,7 +46,7 @@ public class CameraVisualizer : MonoBehaviour
 
     public bool ShowDebugRay = false;
     public Vector2 DebugPixel = new Vector2(0, 0);
-    public int DebugRayMultiplier = 1;
+    public float DebugRayMultiplier = 1;
 
     public bool ShowAntiAliasing = false;
     public int NumAASamples = 0;
